Move Preprocessor into "script" sub-directory
Basically, such preprocessor is a kind of primitive scripting, and it seems better place to store it at "script" sub-directory
