+++ CMakeLists.txt
@@ -624,7 +624,6 @@ set(THEXTECH_SRC
     src/main/level_file.cpp
     src/main/menu_loop.cpp
     src/main/menu_main.cpp
-    src/main/msg_preprocessor.cpp
     src/main/screen_pause.cpp
     src/main/screen_connect.cpp
     src/main/screen_quickreconnect.cpp
@@ -653,6 +652,7 @@ set(THEXTECH_SRC
     src/graphics/gfx_keyhole.cpp
     src/control/duplicate.cpp
     src/control/controls.cpp
+    src/script/msg_preprocessor.cpp
     src/change_res.cpp
     src/effect.cpp
     src/collision.cpp
